refactor: move CalendarScreen to new folder for booking
